dynamicoverlay: use size_t for p_params->i_id
The id for the overlays are necesarily unsigned, and are used as an
array subscript, and then compared against the size of the array, so
using size_t directly as an id ensure we don't have sign warnings.

+++ modules/spu/dynamicoverlay/dynamicoverlay.h
@@ -55,7 +55,7 @@ char *BufferGetToken( buffer_t *p_buffer );
 /** struct commandparams_t - command params structure */
 typedef struct commandparams_t
 {
-    int32_t i_id;       /*< overlay id */
+    size_t  i_id;       /*< overlay id */
     int32_t i_shmid;    /*< shared memory identifier */
 
     vlc_fourcc_t fourcc;/*< chroma */
